requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4

# LangGraph + LangChain core

langgraph==0.0.40
langchain==0.1.16
langchain-core==0.1.46

# Groq LLM

groq==0.5.0

# Database (PostgreSQL)

psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# Utilities

python-dotenv==1.0.1
requests==2.31.0
